# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/RegisterResult"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "fullName",
                            "message": "fullName không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Lỗi, error.code thuộc: AUTH_EMAIL_ALREADY_EXISTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "AUTH_EMAIL_ALREADY_EXISTS": {
                    "summary": "AUTH_EMAIL_ALREADY_EXISTS: Email đã được sử dụng",
                    "value": {
                      "success": false,
                      "status": 409,
                      "error": {
                        "code": "AUTH_EMAIL_ALREADY_EXISTS",
                        "message": "Email đã được sử dụng"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Lỗi, error.code thuộc: TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "TOO_MANY_REQUESTS": {
                    "summary": "TOO_MANY_REQUESTS: Bạn thao tác quá nhanh, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn thao tác quá nhanh, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Đăng ký tài khoản học viên và gửi email xác minh",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: AUTH_TOKEN_INVALID, VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "AUTH_TOKEN_INVALID": {
                    "summary": "AUTH_TOKEN_INVALID: Liên kết không hợp lệ hoặc đã hết hạn",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "AUTH_TOKEN_INVALID",
                        "message": "Liên kết không hợp lệ hoặc đã hết hạn"
                      }
                    }
                  },
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "token",
                            "message": "token không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Lỗi, error.code thuộc: TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "TOO_MANY_REQUESTS": {
                    "summary": "TOO_MANY_REQUESTS: Bạn thao tác quá nhanh, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn thao tác quá nhanh, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Xác minh email bằng token trong liên kết, kích hoạt tài khoản",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "description": "Luôn trả thành công để không lộ email nào đã đăng ký; tối đa 3 lần mỗi 15 phút.",
        "operationId": "resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "email",
                            "message": "email không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Lỗi, error.code thuộc: TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "TOO_MANY_REQUESTS": {
                    "summary": "TOO_MANY_REQUESTS: Bạn thao tác quá nhanh, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn thao tác quá nhanh, vui lòng thử lại sau"
                      }
                    }
                  },
                  "TOO_MANY_REQUESTS_2": {
                    "summary": "TOO_MANY_REQUESTS: Bạn đã yêu cầu quá nhiều lần, vui lòng thử lại sau 15 phút",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn đã yêu cầu quá nhiều lần, vui lòng thử lại sau 15 phút"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Gửi lại email xác minh",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "Luôn trả thành công để không lộ email nào đã đăng ký; tối đa 3 lần mỗi 15 phút.",
        "operationId": "forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "email",
                            "message": "email không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Lỗi, error.code thuộc: TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "TOO_MANY_REQUESTS": {
                    "summary": "TOO_MANY_REQUESTS: Bạn thao tác quá nhanh, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn thao tác quá nhanh, vui lòng thử lại sau"
                      }
                    }
                  },
                  "TOO_MANY_REQUESTS_2": {
                    "summary": "TOO_MANY_REQUESTS: Bạn đã yêu cầu quá nhiều lần, vui lòng thử lại sau 15 phút",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn đã yêu cầu quá nhiều lần, vui lòng thử lại sau 15 phút"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Yêu cầu đặt lại mật khẩu, gửi liên kết qua email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password/validate": {
      "get": {
        "operationId": "validateResetToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Token trong liên kết đặt lại mật khẩu (phần sau \"token=\")",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "status",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi, error.code thuộc: AUTH_TOKEN_INVALID, VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "AUTH_TOKEN_INVALID": {
                    "summary": "AUTH_TOKEN_INVALID: Liên kết không hợp lệ hoặc đã hết hạn",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "AUTH_TOKEN_INVALID",
                        "message": "Liên kết không hợp lệ hoặc đã hết hạn"
                      }
                    }
                  },
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR: Dữ liệu gửi lên không hợp lệ",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Dữ liệu gửi lên không hợp lệ",
                        "details": [
                          {
                            "field": "token",
                            "message": "token không hợp lệ"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Lỗi, error.code thuộc: TOO_MANY_REQUESTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "TOO_MANY_REQUESTS": {
                    "summary": "TOO_MANY_REQUESTS: Bạn thao tác quá nhanh, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 429,
                      "error": {
                        "code": "TOO_MANY_REQUESTS",
                        "message": "Bạn thao tác quá nhanh, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi, error.code thuộc: INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR: Đã có lỗi xảy ra, vui lòng thử lại sau",
                    "value": {
                      "success": false,
                      "status": 500,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Đã có lỗi xảy ra, vui lòng thử lại sau"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Kiểm tra liên kết đặt lại mật khẩu còn hiệu lực (không hủy token)",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart English Exam System API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "status": {
            "type": "integer",
            "example": 400,
            "description": "Luôn trùng HTTP status code"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "status",
          "error"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING_VERIFICATION",
          "PENDING_APPROVAL",
          "ACTIVE",
          "LOCKED"
        ]
      },
      "RegisterResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "hocvien@gmail.com"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "status"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn A",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "example": "hocvien@gmail.com",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "example": "MatKhau123",
            "minLength": 8,
            "maxLength": 72,
            "description": "Mật khẩu dài 8-72 ký tự, gồm ít nhất một chữ hoa, một chữ thường và một chữ số"
          },
          "confirmPassword": {
            "type": "string",
            "example": "MatKhau123"
          },
          "acceptTerms": {
            "type": "boolean",
            "example": true,
            "description": "Đồng ý Điều khoản sử dụng và Chính sách xử lý dữ liệu cá nhân"
          }
        },
        "required": [
          "fullName",
          "email",
          "password",
          "confirmPassword",
          "acceptTerms"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token trong liên kết xác minh (phần sau \"token=\")",
            "example": "q3Zb1k0yXo9mQ2...43 ký tự"
          }
        },
        "required": [
          "token"
        ]
      },
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "hocvien@gmail.com",
            "maxLength": 255
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}