# backend/src/modules/auth/auth.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  HttpStatus,
  Post,
  Query,
  Req,
  Res,
  UseGuards,
} from '@nestjs/common';
import { Throttle, ThrottlerGuard } from '@nestjs/throttler';
import { ConfigService } from '@nestjs/config';
import { ApiCookieAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
import type { CookieOptions, Request, Response } from 'express';
import { IsPublic } from '../../common/decorators/public.decorator.js';
import { ApiEnvelope, ApiNullEnvelope } from '../../common/swagger/api-envelope.decorator.js';
import { ApiErrors } from '../../common/swagger/api-errors.decorator.js';
import type { ApiErrorSpec } from '../../common/swagger/api-errors.decorator.js';
import { AppException } from '../../common/errors/app.exception.js';
import { ErrorCode } from '../../common/errors/error-codes.js';
import type { Env } from '../../config/env.schema.js';
import { AUTH_THROTTLE, REFRESH_COOKIE_NAME, REFRESH_COOKIE_PATH } from './auth.constants.js';
import { AuthService } from './auth.service.js';
import { AuthSession, RegisterResult } from './auth.types.js';
import type { RequestMeta } from './auth.types.js';
import { LoginDto } from './dto/login.dto.js';
import { RegisterDto } from './dto/register.dto.js';
import { VerifyEmailDto } from './dto/verify-email.dto.js';
import { EmailDto } from './dto/email.dto.js';
import { ResetPasswordDto, ResetPasswordTokenDto } from './dto/reset-password.dto.js';

// Gửi mail (gửi lại xác minh, quên mật khẩu): 429 do giới hạn theo IP hoặc theo email
const EMAIL_REQUEST_ERRORS: ApiErrorSpec[] = [
  ErrorCode.TOO_MANY_REQUESTS,
  {
    code: ErrorCode.TOO_MANY_REQUESTS,
    message: 'Bạn đã yêu cầu quá nhiều lần, vui lòng thử lại sau 15 phút',
  },
];

@ApiTags('Auth')
@IsPublic()
// Giới hạn theo IP cho mọi route auth (mức default); route nhạy cảm ghi đè bằng @Throttle
@UseGuards(ThrottlerGuard)
@Controller('auth')
export class AuthController {
  private readonly refreshCookieOptions: CookieOptions;
  private readonly refreshCookieMaxAge: number;

  constructor(
    private readonly authService: AuthService,
    config: ConfigService<Env, true>,
  ) {
    this.refreshCookieOptions = {
      httpOnly: true, // JavaScript phía trình duyệt không đọc được, chống XSS lấy token
      sameSite: 'strict', // không gửi kèm request từ trang khác, chống CSRF
      path: REFRESH_COOKIE_PATH, // chỉ gửi kèm các route /api/v1/auth/*
      // Dev chạy http nên chỉ bật Secure ở production
      secure: config.get('NODE_ENV', { infer: true }) === 'production',
    };
    this.refreshCookieMaxAge = config.get('JWT_REFRESH_TTL', { infer: true }) * 1000;
  }

  @Post('register')
  @Throttle({ default: AUTH_THROTTLE.sensitive })
  @ApiOperation({ summary: 'Đăng ký tài khoản học viên và gửi email xác minh' })
  @ApiEnvelope(RegisterResult, { status: HttpStatus.CREATED })
  @ApiErrors(ErrorCode.AUTH_EMAIL_ALREADY_EXISTS, ErrorCode.TOO_MANY_REQUESTS)
  register(@Body() dto: RegisterDto): Promise<RegisterResult> {
    return this.authService.register(dto);
  }

  @Post('verify-email')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Xác minh email bằng token trong liên kết, kích hoạt tài khoản' })
  @ApiNullEnvelope()
  @ApiErrors(ErrorCode.AUTH_TOKEN_INVALID, ErrorCode.TOO_MANY_REQUESTS)
  verifyEmail(@Body() dto: VerifyEmailDto): Promise<void> {
    return this.authService.verifyEmail(dto.token);
  }

  @Post('resend-verification')
  @Throttle({ default: AUTH_THROTTLE.sensitive })
  @HttpCode(HttpStatus.OK)
  @ApiOperation({
    summary: 'Gửi lại email xác minh',
    description: 'Luôn trả thành công để không lộ email nào đã đăng ký; tối đa 3 lần mỗi 15 phút.',
  })
  @ApiNullEnvelope()
  @ApiErrors(...EMAIL_REQUEST_ERRORS)
  resendVerification(@Body() dto: EmailDto): Promise<void> {
    return this.authService.resendVerification(dto.email);
  }

  @Post('forgot-password')
  @Throttle({ default: AUTH_THROTTLE.sensitive })
  @HttpCode(HttpStatus.OK)
  @ApiOperation({
    summary: 'Yêu cầu đặt lại mật khẩu, gửi liên kết qua email',
    description: 'Luôn trả thành công để không lộ email nào đã đăng ký; tối đa 3 lần mỗi 15 phút.',
  })
  @ApiNullEnvelope()
  @ApiErrors(...EMAIL_REQUEST_ERRORS)
  forgotPassword(@Body() dto: EmailDto): Promise<void> {
    return this.authService.forgotPassword(dto.email);
  }

  @Get('reset-password/validate')
  @ApiOperation({ summary: 'Kiểm tra liên kết đặt lại mật khẩu còn hiệu lực (không hủy token)' })
  @ApiNullEnvelope()
  @ApiErrors(ErrorCode.AUTH_TOKEN_INVALID, ErrorCode.TOO_MANY_REQUESTS)
  validateResetToken(@Query() dto: ResetPasswordTokenDto): Promise<void> {
    return this.authService.validateResetToken(dto.token);
  }

  @Post('reset-password')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Đặt mật khẩu mới bằng token, đăng xuất mọi thiết bị' })
  @ApiNullEnvelope()
  @ApiErrors(ErrorCode.AUTH_TOKEN_INVALID, ErrorCode.TOO_MANY_REQUESTS)
  resetPassword(@Body() dto: ResetPasswordDto): Promise<void> {
    return this.authService.resetPassword(dto);
  }

  @Post('login')
  @Throttle({ default: AUTH_THROTTLE.login })
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Đăng nhập bằng email và mật khẩu (dùng chung cho mọi vai trò)' })
  @ApiEnvelope(AuthSession)
  @ApiErrors(
    ErrorCode.AUTH_INVALID_CREDENTIALS,
    ErrorCode.AUTH_EMAIL_NOT_VERIFIED,
    ErrorCode.AUTH_ACCOUNT_PENDING_APPROVAL,
    ErrorCode.AUTH_ACCOUNT_LOCKED,
    {
      code: ErrorCode.AUTH_TOO_MANY_LOGIN_ATTEMPTS,
      details: { retryAfterSeconds: 900 },
    },
    ErrorCode.TOO_MANY_REQUESTS,
  )
  async login(
    @Body() dto: LoginDto,
    @Req() req: Request,
    @Res({ passthrough: true }) res: Response,
  ): Promise<AuthSession> {
    const { refreshToken, ...session } = await this.authService.login(dto, this.requestMeta(req));
    this.setRefreshCookie(res, refreshToken);
    return session;
  }

  @Post('refresh')
  @HttpCode(HttpStatus.OK)
  @ApiCookieAuth(REFRESH_COOKIE_NAME)
  @ApiOperation({ summary: 'Cấp lại access token, xoay vòng refresh token trong cookie' })
  @ApiEnvelope(AuthSession)
  @ApiErrors(ErrorCode.AUTH_REFRESH_TOKEN_INVALID, ErrorCode.TOO_MANY_REQUESTS)
  async refresh(
    @Req() req: Request,
    @Res({ passthrough: true }) res: Response,
  ): Promise<AuthSession> {
    try {
      const { refreshToken, ...session } = await this.authService.refresh(
        this.readRefreshCookie(req),
      );
      this.setRefreshCookie(res, refreshToken);
      return session;
    } catch (error) {
      // Phiên không cứu được nữa: xóa cookie để trình duyệt thôi gửi token hỏng.
      // Lỗi khác (Redis/DB tạm lỗi) thì giữ cookie để người dùng thử lại.
      if (error instanceof AppException && error.code === ErrorCode.AUTH_REFRESH_TOKEN_INVALID) {
        this.clearRefreshCookie(res);
      }
      throw error;
    }
  }

  @Post('logout')
  @HttpCode(HttpStatus.OK)
  @ApiCookieAuth(REFRESH_COOKIE_NAME)
  @ApiOperation({ summary: 'Đăng xuất, thu hồi refresh token của phiên hiện tại' })
  @ApiNullEnvelope()
  @ApiErrors(ErrorCode.TOO_MANY_REQUESTS)
  async logout(@Req() req: Request, @Res({ passthrough: true }) res: Response): Promise<void> {
    await this.authService.logout(this.readRefreshCookie(req));
    this.clearRefreshCookie(res);
  }

  private setRefreshCookie(res: Response, token: string): void {
    res.cookie(REFRESH_COOKIE_NAME, token, {
      ...this.refreshCookieOptions,
      maxAge: this.refreshCookieMaxAge,
    });
  }

  // Phải truyền cùng path, sameSite... như lúc đặt thì trình duyệt mới xóa đúng cookie
  private clearRefreshCookie(res: Response): void {
    res.clearCookie(REFRESH_COOKIE_NAME, this.refreshCookieOptions);
  }

  private readRefreshCookie(req: Request): string | undefined {
    const cookies = req.cookies as Record<string, string | undefined> | undefined;
    return cookies?.[REFRESH_COOKIE_NAME];
  }

  private requestMeta(req: Request): RequestMeta {
    return { ipAddress: req.ip, userAgent: req.get('user-agent') };
  }
}